api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
authors = [
    {name = "moizahmedshaikh",email = "moizaman905@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11, <3.13"
dependencies = [
    "fastapi (>=0.123.0,<0.124.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "python-dotenv (>=1.0.1)",
    "openai-agents (>=0.1.0,<0.2.0)",
    "fastembed (>=0.2.7,<0.3.0)",
    "langchain-google-genai>=1.0.5",
    "langchain>=0.2.1",
    "langchain-qdrant>=0.1.1",
    "unstructured[md]>=0.14.5",
    "qdrant-client (>=1.16.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
